requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2
django-allauth==0.57.0
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-redis==5.3.0
django-simple-captcha==0.5.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
memory-profiler==0.61.0
mysql-connector-python==8.2.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.21.12
psutil==5.9.5
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
